% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{connect}
\alias{connect}
\title{Connect to an OmniSci database}
\usage{
connect(host, port, username, passwd, db_name)
}
\arguments{
\item{host}{host}

\item{port}{port}

\item{username}{username}

\item{passwd}{passwd}

\item{db_name}{db_name}
}
\description{
Connect to an OmniSci database
}
