% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{connect_binary}
\alias{connect_binary}
\title{Connect to an OmniSci database}
\usage{
connect_binary(host, port, username, password, dbname)
}
\arguments{
\item{host}{host}

\item{port}{port}

\item{username}{username}

\item{password}{password}

\item{dbname}{dbname}
}
\value{
List
}
\description{
TBD description
}
\details{
TBD Details
}
\examples{
\dontrun{
  
  conn <- connect("localhost", 6274, "admin", "HyperInteractive", "omnisci")
  
}
}
