% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{disconnect}
\alias{disconnect}
\title{THRIFT: Disconnect from an OmniSci database}
\usage{
disconnect(conn)
}
\arguments{
\item{conn}{conn}
}
\value{
Nothing
}
\description{
TBD description
}
\details{
TBD Details
}
\examples{
\dontrun{
  
  disconnect(conn)
  
}
}
