% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_databases}
\alias{get_databases}
\title{Get databases list}
\usage{
get_databases(conn)
}
\arguments{
\item{conn}{conn}
}
\value{
List(TDBInfo)
}
\description{
TBD description
}
\details{
TBD Details
}
\examples{
\dontrun{

gdb <- get_databases(conn)

} 
}
