% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_hardware_info}
\alias{get_hardware_info}
\title{THRIFT: Get hardware info}
\usage{
get_hardware_info(conn)
}
\arguments{
\item{conn}{conn}
}
\value{
List(TClusterHardwareInfo)
}
\description{
TBD description
}
\details{
TBD Details
}
\examples{
\dontrun{

ghi <- get_hardware_info(conn)

} 
}
