% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_server_status}
\alias{get_server_status}
\title{THRIFT: Get status of OmniSci server}
\usage{
get_server_status(conn)
}
\arguments{
\item{conn}{conn}
}
\value{
List (TServerStatus)
}
\description{
TBD description
}
\details{
TBD Details
}
\examples{
\dontrun{

gss <- get_server_status(conn)

} 
}
