% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_session_info}
\alias{get_session_info}
\title{Get session information}
\usage{
get_session_info(conn)
}
\arguments{
\item{conn}{conn}
}
\value{
List(TSessionInfo)
}
\description{
TBD description
}
\details{
TBD Details
}
\examples{
\dontrun{

gsi <- get_session_info(conn)

} 
}
