% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_tables}
\alias{get_tables}
\title{THRIFT: Get tables in current database}
\usage{
get_tables(conn)
}
\arguments{
\item{conn}{conn}
}
\value{
List(Character)
}
\description{
TBD description
}
\details{
TBD Details
}
\examples{
\dontrun{

gt <- get_tables(conn)

} 
}
