% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_tables_meta}
\alias{get_tables_meta}
\title{THRIFT: Get metadata of tables in current database}
\usage{
get_tables_meta(conn)
}
\arguments{
\item{conn}{conn}
}
\value{
List(TTableMeta)
}
\description{
TBD description
}
\details{
TBD Details
}
\examples{
\dontrun{

gtm <- get_tables_meta(conn)

} 
}
