% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_users}
\alias{get_users}
\title{THRIFT: Get users in current database}
\usage{
get_users(conn)
}
\arguments{
\item{conn}{conn}
}
\value{
List(Character)
}
\description{
TBD description
}
\details{
TBD Details
}
\examples{
\dontrun{

gu <- get_users(conn)

} 
}
