% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_version}
\alias{get_version}
\title{Get version of OmniSci server}
\usage{
get_version(conn)
}
\arguments{
\item{conn}{conn}
}
\value{
Character
}
\description{
TBD description
}
\details{
TBD Details
}
\examples{
\dontrun{

gv <- get_version(conn)

} 
}
