% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_views}
\alias{get_views}
\title{Get views in current database}
\usage{
get_views(conn)
}
\arguments{
\item{conn}{conn}
}
\value{
List(Character)
}
\description{
TBD description
}
\details{
TBD Details
}
\examples{
\dontrun{

gvs <- get_views(conn)

} 
}
