% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sql_execute}
\alias{sql_execute}
\title{Execute SQL statement}
\usage{
sql_execute(conn, query, first_n = -1L, at_most_n = -1L)
}
\arguments{
\item{conn}{conn}

\item{query}{query}

\item{first_n}{first_n}

\item{at_most_n}{at_most_n}
}
\value{
List(TQueryResult)
}
\description{
TBD description
}
\details{
TBD Details
}
\examples{
\dontrun{

result <- sql_execute(conn, "select * from omnisci_states")

} 
}
