% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{switch_database}
\alias{switch_database}
\title{THRIFT: Switch current database}
\usage{
switch_database(conn, dbname)
}
\arguments{
\item{conn}{conn}

\item{dbname}{dbname}
}
\value{
Nothing
}
\description{
TBD description
}
\details{
TBD Details
}
\examples{
\dontrun{

switch_database(conn, "database2")

} 
}
