% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_table_details}
\alias{get_table_details}
\title{Get details of the specified database table}
\usage{
get_table_details(conn, table_name)
}
\arguments{
\item{conn}{conn}

\item{table_name}{table_name}
}
\value{
List (TTableDetails)
}
\description{
TBD description
}
\details{
TBD Details
}
\examples{
\dontrun{

tbl_details <- get_table_details(conn, "omnisci_states")

} 
}
